Switch section page to show related simp and trad on same row
This is more in line with Chloe's design, a deviation from the legacy site.
It's partially to simplify the layout, partially to simplify code.
It's easier to repurpose the table for section editing if it's already organized
by section row rather than by individual vocab.

The main downside that I see is that users don't have as fine-grained control
over which vocabs get banned or unbanned, but I think such controls
will be in a vocab modal, and so we can have the section page do sweeping changes.

The other downside is that, for users studying only simp or only trad, this
is more information than they need. But I think this is outweighed by the
simplicity from having the page more the same from user to user.