Add spoof buttons, policies, cancel button to sub page
Testing all the different combinations of subscription settings
is complicated, so I added buttons for admins to use to see what
the page looks like under various subscriptions and clients, and
with discounts or vacations on or off. Also did some further work on
the template and some edge cases. Cancel button is not yet working.